requirements.txt
gradio==3.39.0
pillow==10.4.0
numpy==1.23.5
tensorflow-cpu==2.12.0
h5py==3.13.0
opencv-python-headless==4.11.0.86
gdown==5.2.0
